# local/rest_api_gcbm/static/swagger.json
{
...
    "info": {
        "title": "FLINT-GCBM Rest Api",
        "version": "1.0"
    },
    "paths": {
...
        "/help/{arg}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "description": "Help info about named section. Pass all to get all info",
                        "in": "path",
                        "name": "arg",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Help"
                    }
                },
                "summary": "Get Help Section",
                "tags": [
                    "help"
                ]
            }
        },
...
    },
    "swagger": "2.0"
}